chore: refactor to only calculate chunks once
